backend/requirements.txt
django==4.2.13
djangorestframework==3.15.2
coreapi-cli==1.0.9
django-cors-headers==4.4.0
psycopg2-binary
wheel==0.43.0
django-environ
pillow
easy-thumbnails
pandas==2.0.3
catboost==1.2.7
numpy==1.24.3
